requirements.txt
allensdk==0.14.4
argh==0.26.2
asn1crypto==0.24.0
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
bitstring==3.1.5
cachetools==2.0.1
cbor==1.0.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
crossbar==18.3.1
cryptography==2.2.2
cycler==0.10.0
dask==0.17.2
decorator==4.2.1
enum34==1.1.6
future==0.16.0
h2==3.0.1
h5py==2.7.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==18.0.0
idna==2.5
incremental==17.5.0
ipaddress==1.0.19
Jinja2==2.10
kiwisolver==1.0.1
klein==17.10.0
llvmlite==0.22.0
lmdb==0.94
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
netaddr==0.7.19
netCDF4==1.3.1
networkx==2.1
numba==0.37.0
numpy==1.13.0
pandas==0.22.0
pathtools==0.1.2
patsy==0.5.0
Pillow==5.1.0
priority==1.3.0
psutil==5.4.3
py-ubjson==0.11.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pynrrd==0.2.4
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyQRCode==1.2.1
python-dateutil==2.7.2
PyTrie==0.3
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-toolbelt==0.8.0
scikit-image==0.13.1
scipy==1.0.1
sdnotify==0.3.2
service-identity==17.0.0
setproctitle==1.1.10
shutilwhich==1.1.0
simplejson==3.13.2
six==1.11.0
sortedcontainers==1.5.9
statsmodels==0.8.0
toolz==0.9.0
treq==17.8.0
Twisted==17.9.0
txaio==2.10.0
txredisapi==1.4.4
txtorcon==0.20.0
u-msgpack-python==2.5.0
urllib3==1.22
watchdog==0.8.3
Werkzeug==0.14.1
xarray==0.10.2
zope.interface==4.4.3
